Load dashboard AI modals on demand
