+++ tests/test-modal-lifecycle-integrations.js
@@ -14,7 +14,7 @@ const chatSummariesSrc = fs.readFileSync(path.join(root, 'js/chat-summaries.js')
 const contextCardsSrc = fs.readFileSync(path.join(root, 'js/context-cards.js'), 'utf8');
 const contextLifestyleSrc = fs.readFileSync(path.join(root, 'js/context-card-lifestyle-editors-impl.js'), 'utf8');
 const contextMedicalSrc = fs.readFileSync(path.join(root, 'js/context-card-medical-history-editor-impl.js'), 'utf8');
-const dashboardAiSrc = fs.readFileSync(path.join(root, 'js/context-card-dashboard-ai.js'), 'utf8');
+const dashboardAiSrc = fs.readFileSync(path.join(root, 'js/context-card-dashboard-ai-impl.js'), 'utf8');
 const dashboardWidgetControlsSrc = fs.readFileSync(path.join(root, 'js/dashboard-widget-controls.js'), 'utf8');
 const cycleSrc = fs.readFileSync(path.join(root, 'js/cycle.js'), 'utf8');
 const emfSrc = fs.readFileSync(path.join(root, 'js/emf.js'), 'utf8');
