+++ tsconfig.json
@@ -83,6 +83,7 @@
     "js/constants.js",
     "js/context-card-dashboard-ai-actions.js",
     "js/context-card-dashboard-ai.js",
+    "js/context-card-dashboard-ai-impl.js",
     "js/context-card-editor-ui.js",
     "js/context-card-health-dots.js",
     "js/context-card-lifestyle-runtime.js",
